Install to .local and add zef
Using .local is more convenient because it is already in the path.

Package installation via zef is performed via Ansible rather than the
equivalent Makefile target to avoid having to account for the Makefile
during deployment. The number of packages to account for is small
enough for this duplication to not be too terrible.